refactor : 패키지 구조를 바꿔라
- domain, controller패키지 구조로 수정함
