+++ src/main/java/com/codessquad/qna/web/controller/QuestionsController.java
@@ -1,8 +1,10 @@
-package com.codessquad.qna.web.questions;
+package com.codessquad.qna.web.controller;
 
+import com.codessquad.qna.web.domain.Question;
+import com.codessquad.qna.web.domain.QuestionRepository;
 import com.codessquad.qna.web.exceptions.auth.UnauthorizedAccessException;
 import com.codessquad.qna.web.exceptions.questions.QuestionNotFoundException;
-import com.codessquad.qna.web.users.User;
+import com.codessquad.qna.web.domain.User;
 import com.codessquad.qna.web.utils.SessionUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
