+++ src/main/java/com/codessquad/qna/web/domain/AnswersRepository.java
@@ -1,5 +1,6 @@
-package com.codessquad.qna.web.answers;
+package com.codessquad.qna.web.domain;
 
+import com.codessquad.qna.web.domain.Answer;
 import org.springframework.data.repository.CrudRepository;
 
 public interface AnswersRepository extends CrudRepository<Answer, Long> {
