+++ src/main/java/com/codessquad/qna/web/domain/QuestionRepository.java
@@ -1,5 +1,6 @@
-package com.codessquad.qna.web.questions;
+package com.codessquad.qna.web.domain;
 
+import com.codessquad.qna.web.domain.Question;
 import org.springframework.data.repository.CrudRepository;
 
 public interface QuestionRepository extends CrudRepository<Question, Long> {
